Fixed layout + fixed broken ssr for desktops
***
frontend
