+++ frontend/app/lib/components/adaptiveContainer.tsx
@@ -1,82 +1,27 @@
-'use client';
-
-import React, { ReactElement, useMemo } from 'react';
-import { Tab, Tabs } from '@heroui/tabs';
-import useAdaptiveParams from '@/app/lib/hooks/useAdaptiveParams';
-import { observer } from 'mobx-react-lite';
-import settingsStore from '@/app/stores/settingsStore';
+import React from 'react';
 import clsx from 'clsx';
+import MobileLayout from '@/app/lib/components/mobileLayout';
 
-interface MobileTabProps {
-  key: string;
-  titleKey: string;
+export default function AdaptiveContainer({
+  children,
+  mobileTabs,
+  desktopContainerClassName,
+}: Readonly<{
+  children: React.ReactNode;
+  mobileTabs: {
+    key: string;
+    titleKey: string;
+  }[];
+  desktopContainerClassName?: string;
+}>) {
+  return (
+    <>
+      <div className={clsx('hidden sm:flex', desktopContainerClassName)}>
+        {children}
+      </div>
+      <div className="flex h-full max-h-screen w-full flex-col gap-2 p-4 pb-4 sm:hidden">
+        <MobileLayout mobileTabs={mobileTabs}>{children}</MobileLayout>
+      </div>
+    </>
+  );
 }
-
-const AdaptiveContainer = observer(
-  ({
-    children,
-    mobileTabs,
-    desktopContainerClassName,
-  }: Readonly<{
-    children: React.ReactNode;
-    mobileTabs: MobileTabProps[];
-    desktopContainerClassName?: string;
-  }>) => {
-    const { commonSize } = useAdaptiveParams();
-
-    const childMap = useMemo(() => {
-      const map = new Map<string, ReactElement>();
-
-      const childrenArray = children as any;
-
-      const a = Array.isArray(childrenArray) ? childrenArray : [childrenArray];
-
-      a.forEach((child: any) => {
-        if (!child || typeof child !== 'object') return;
-
-        const key = child.key as string | null;
-        if (!key) return;
-
-        const normalizedKey = key.startsWith('.$')
-          ? key.slice(2)
-          : key.replace(/^\./, '');
-        map.set(normalizedKey, child);
-      });
-
-      return map;
-    }, [children]);
-
-    return (
-      <>
-        <div className={clsx('hidden sm:flex', desktopContainerClassName)}>
-          {children}
-        </div>
-        <div className="flex h-full max-h-screen w-full flex-col gap-2 p-4 pb-4 sm:hidden">
-          <Tabs
-            size={commonSize}
-            variant="bordered"
-            color="primary"
-            aria-label="Tabs"
-            items={mobileTabs}
-            classNames={{
-              tabList: 'w-full rounded-xl',
-              panel: 'p-0 min-h-0 flex-1 font-medium overflow-clip',
-            }}
-          >
-            {(tab: MobileTabProps) => {
-              return (
-                <Tab
-                  key={tab.key}
-                  title={settingsStore.t.main.mobileTitles[tab.titleKey]}
-                >
-                  {childMap.get(tab.key)}
-                </Tab>
-              );
-            }}
-          </Tabs>
-        </div>
-      </>
-    );
-  },
-);
-export default AdaptiveContainer;
